Single-host with structured (continuous) host population

This function runs a single-host transmission chain simulation, with a structured host population (such as spatial features). The simulation stops either at the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected) is passed.

singleContinuous(length.sim, max.infected, init.individuals,
  init.structure, structure.raster, diff.pMove = FALSE,
  timeDep.pMove = FALSE, pMove, param.pMove, diff.sdMove = FALSE,
  timeDep.sdMove = FALSE, sdMove, param.sdMove,
  attracted.by.raster = FALSE, diff.nContact = FALSE,
  timeDep.nContact = FALSE, nContact, param.nContact,
  diff.pTrans = FALSE, timeDep.pTrans = FALSE, pTrans, param.pTrans,
  diff.pExit = FALSE, timeDep.pExit = FALSE, pExit, param.pExit,
  prefix.host = "H", progress.bar = TRUE, print.step = 10)

Arguments

length.sim
specifies the length (in unit of time) over which the simulation should be run.

max.infected
specifies the maximum number of hosts that can be infected in the simulation.

init.individuals
number of initially infected individuals.

init.structure
which state (i.e. location) the initially infected individuals are located.

structure.raster
raster object defining the environmental variable.

diff.pMove
is pMove different between states of the structured population (TRUE/FALSE)

timeDep.pMove
is pMove dependant on the absolute time of the simulation (TRUE/FALSE)

pMove
function that gives the probability of a host moving as a function of time.

param.pMove
parameter names (list of functions) for the pMove.

diff.sdMove
is sdMove dependant on the environmental value (TRUE/FALSE).

timeDep.sdMove
is sdMove dependant on the absolute time of the simulation (TRUE/FALSE).

sdMove
function that gives the distance travelled (based on coordinates); gives the sd value for the brownian motion.

param.sdMove
parameter names (list of functions) for sdMove

attracted.by.raster
should the hosts be attracted by high values in the environmental raster? (TRUE/FALSE)

diff.nContact
is nContact different between states of the structured population (TRUE/FALSE)

timeDep.nContact
is nContact dependant on the absolute time of the simulation (TRUE/FALSE)

nContact
function that gives the number of potential transmission events per unit of time.

param.nContact
parameter names (list of functions) for nContact.

diff.pTrans
is pTrans different between states of the structured population (TRUE/FALSE)

timeDep.pTrans
is pTrans dependant on the absolute time of the simulation (TRUE/FALSE)

pTrans
function that gives the probability of transmit a pathogen as a function of time since infection.

param.pTrans
parameter names (list of functions) for the pTrans.

diff.pExit
is pExit different between states of the structured population (TRUE/FALSE)

timeDep.pExit
is pExit dependant on the absolute time of the simulation (TRUE/FALSE)

pExit
function that gives the probability to exit the simulation for an infected host (either moving out, dying, etc.).

param.pExit
parameter names (list of functions) for the pExit.

prefix.host
character(s) to be used as a prefix for the hosts identification number.

progress.bar
if TRUE, print message on simulation state to screen.

print.step
progress.bar is TRUE, step with which the progress message will be printed.