Dual-host with structured (discrete) host population

This function runs a dual-host transmission chain simulation, with discrete structure features. The simulation stops either at the end of given time (specified by length.sim) or when the number of hosts infected threshold (max.infected) is passed.

dualContinuous(length.sim, max.infected.A, max.infected.B,
  init.individuals.A, init.individuals.B, init.structure.A,
  init.structure.B, structure.raster.A, structure.raster.B, pExit.A,
  param.pExit.A, timeDep.pExit.A = FALSE, diff.pExit.A = FALSE,
  hostCount.pExit.A = FALSE, pMove.A, param.pMove.A,
  timeDep.pMove.A = FALSE, diff.pMove.A = FALSE,
  hostCount.pMove.A = FALSE, sdMove.A, param.sdMove.A,
  diff.sdMove.A = FALSE, timeDep.sdMove.A = FALSE,
  hostCount.sdMove.A = FALSE, attracted.by.raster.A = FALSE,
  nContact.A, param.nContact.A, timeDep.nContact.A = FALSE,
  diff.nContact.A = FALSE, hostCount.nContact.A = FALSE, pTrans.A,
  param.pTrans.A, timeDep.pTrans.A = FALSE, diff.pTrans.A = FALSE,
  hostCount.pTrans.A = FALSE, prefix.host.A = "H", pExit.B,
  param.pExit.B, timeDep.pExit.B = FALSE, diff.pExit.B = FALSE,
  hostCount.pExit.B = FALSE, pMove.B, param.pMove.B,
  timeDep.pMove.B = FALSE, diff.pMove.B = FALSE,
  hostCount.pMove.B = FALSE, sdMove.B, param.sdMove.B,
  diff.sdMove.B = FALSE, timeDep.sdMove.B = FALSE,
  hostCount.sdMove.B = FALSE, attracted.by.raster.B = FALSE,
  nContact.B, param.nContact.B, timeDep.nContact.B = FALSE,
  diff.nContact.B = FALSE, hostCount.nContact.B = FALSE, pTrans.B,
  param.pTrans.B, timeDep.pTrans.B = FALSE, diff.pTrans.B = FALSE,
  hostCount.pTrans.B = FALSE, prefix.host.B = "V",
  print.progress = TRUE, print.step = 10)

Arguments

length.sim
specifies the length (in unit of time) over which the simulation should be run.

max.infected.A
specifies the maximum number of hosts A that can be infected in the simulation.

max.infected.B
specifies the maximum number of hosts B that can be infected in the simulation.

init.individuals.A
number of initially infected individuals (hosts A).

init.individuals.B
number of initially infected individuals (hosts B).

init.structure.A
which state (e.g. location) the initially infected individuals of host A are located. (NA if init.individual.A is 0)

init.structure.B
which state (e.g. location) the initially infected individuals of host B are located. (NA if init.individual.B is 0)

structure.raster.A
raster object defining the environmental variable for host A.

structure.raster.B
raster object defining the environmental variable for host B.

pExit.A
function that gives the probability to exit the simulation for an infected host A (either moving out, dying, etc.).

param.pExit.A
parameter names (list of functions) for the pExit for host A.

timeDep.pExit.A
is pExit dependant on the absolute time of the simulation (TRUE/FALSE) for host A.

diff.pExit.A
is pExit different between states of the structured population (TRUE/FALSE) for host A.

hostCount.pExit.A
does pExit varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pExit.A should be TRUE.

pMove.A
function that gives the probability of a host moving as a function of time for host A.

param.pMove.A
parameter names (list of functions) for the pMove for host A.

timeDep.pMove.A
is pMove dependant on the absolute time of the simulation (TRUE/FALSE) for host A.

diff.pMove.A
is pMove different between states of the structured population (TRUE/FALSE) for host A.

hostCount.pMove.A
does pMove varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pMove.A should be TRUE.

sdMove.A
function that gives the distance travelled (based on coordinates); gives the sd value for the brownian motion for host A.

param.sdMove.A
parameter names (list of functions) for sdMove for host A

diff.sdMove.A
is sdMove dependant on the environmental value (TRUE/FALSE) for host A.

timeDep.sdMove.A
is sdMove dependant on the absolute time of the simulation (TRUE/FALSE) for host A.

hostCount.sdMove.A
does sdMove varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pMove.A should be TRUE.

attracted.by.raster.A
should the hosts A be attracted by high values in the environmental raster? (TRUE/FALSE)

nContact.A
function that gives the number of potential transmission events per unit of time for host A.

param.nContact.A
parameter names (list of functions) for param.nContact for host A.

timeDep.nContact.A
is nContact dependant on the absolute time of the simulation (TRUE/FALSE) for host A.

diff.nContact.A
is nContact different between states of the structured population (TRUE/FALSE) for host A.

hostCount.nContact.A
does nContact varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.nContact.A should be TRUE.

pTrans.A
function that gives the probability of transmit a pathogen as a function of time since infection for host A.

param.pTrans.A
parameter names (list of functions) for the pExit for host A.

timeDep.pTrans.A
is pTrans dependant on the absolute time of the simulation (TRUE/FALSE) for host A.

diff.pTrans.A
is pTrans different between states of the structured population (TRUE/FALSE) for host A.

hostCount.pTrans.A
does pTrans varies with the host count (of either host A or B) in the state for host A? (TRUE/FALSE); diff.pTrans.A should be TRUE.

prefix.host.A
character(s) to be used as a prefix for the host A identification number.

pExit.B
function that gives the probability to exit the simulation for an infected host B (either moving out, dying, etc.).

param.pExit.B
parameter names (list of functions) for the pExit for host B.

timeDep.pExit.B
is pExit dependant on the absolute time of the simulation (TRUE/FALSE) for host B.

diff.pExit.B
is pExit different between states of the structured population (TRUE/FALSE) for host B.

hostCount.pExit.B
does pExit varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pExit.B should be TRUE.

pMove.B
function that gives the probability of a host moving as a function of time for host B.

param.pMove.B
parameter names (list of functions) for the pMove for host B.

timeDep.pMove.B
is pMove dependant on the absolute time of the simulation (TRUE/FALSE) for host B.

diff.pMove.B
is pMove different between states of the structured population (TRUE/FALSE) for host B.

hostCount.pMove.B
does pMove varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pMove.B should be TRUE.

sdMove.B
function that gives the distance travelled (based on coordinates); gives the sd value for the brownian motion for host A.

param.sdMove.B
parameter names (list of functions) for sdMove for host A

diff.sdMove.B
is sdMove dependant on the environmental value (TRUE/FALSE) for host A.

timeDep.sdMove.B
is sdMove dependant on the absolute time of the simulation (TRUE/FALSE) for host A.

hostCount.sdMove.B
does sdMove varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pMove.B should be TRUE.

attracted.by.raster.B
should the hosts A be attracted by high values in the environmental raster? (TRUE/FALSE)

nContact.B
function that gives the number of potential transmission events per unit of time for host B.

param.nContact.B
parameter names (list of functions) for param.nContact for host B.

timeDep.nContact.B
is nContact dependant on the absolute time of the simulation (TRUE/FALSE) for host B.

diff.nContact.B
is nContact different between states of the structured population (TRUE/FALSE) for host B.

hostCount.nContact.B
does nContact varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.nContact.B should be TRUE.

pTrans.B
function that gives the probability of transmit a pathogen as a function of time since infection for host B.

param.pTrans.B
parameter names (list of functions) for the pExit for host B.

timeDep.pTrans.B
is pTrans dependant on the absolute time of the simulation (TRUE/FALSE) for host B.

diff.pTrans.B
is pTrans different between states of the structured population (TRUE/FALSE) for host B.

hostCount.pTrans.B
does pTrans varies with the host count (of either host A or B) in the state for host B? (TRUE/FALSE); diff.pTrans.B should be TRUE.

prefix.host.B
character(s) to be used as a prefix for the host B identification number.

print.progress
if TRUE, displays a progress bar (current time/length.sim).

print.step
print.progress is TRUE, step with which the progress message will be printed.